The boilerplate has been restructures and modularized
